adjusted chat navbars
